% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_2_filename.R
\name{knit_2_filename}
\alias{knit_2_filename}
\title{Custom Knit function for RStudio
Compiles the current file to the input filename}
\usage{
knit_2_filename(input, ...)
}
\description{
Custom Knit function for RStudio
Compiles the current file to the input filename
}
